# vacation-pay-calculator/src/main/resources/api-docs.yaml
openapi: 3.0.1
info:
  title: Vacation Pay Calculator
  description: App that helps people to calculate their vacation pay
  contact:
    name: Ponomareva Anna
    url: https://github.com/Anna-Votinova
    email: anyvotinova@yandex.ru
  version: 0.0.1-SNAPSHOT
servers:
- url: http://localhost:8080
  description: Generated server url
tags:
- name: Расчет отпускных
  description: Позволяет расчитать отпускные
paths:
  /calculate:
    get:
      tags:
      - Расчет отпускных
      summary: Получение суммы отпускных
      description: "Позволяет получить сумму отпускных на основе введенных данных\
        : если не будут введены точные даты начала и окончания отпуска, калькулято\
        р посчитает отпускные по количеству введенных дней отпуска"
      operationId: getVacationPay
      parameters:
      - name: salary
        in: query
        description: Средняя зарплата за месяц
        required: true
        schema:
          type: number
      - name: days
        in: query
        description: Количество дней отпуска
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: from
        in: query
        description: Дата начала отпуска
        required: false
        schema:
          type: string
          format: date
      - name: to
        in: query
        description: Дата окончания отпуска
        required: false
        schema:
          type: string
          format: date
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                oneOf:
                - $ref: '#/components/schemas/ErrorResponse'
                - $ref: '#/components/schemas/ValidationErrorResponse'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: number
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
    ValidationErrorResponse:
      type: object
      properties:
        violations:
          type: array
          items:
            $ref: '#/components/schemas/Violation'
    Violation:
      type: object
      properties:
        fieldName:
          type: string
        message:
          type: string